// src/features/panels/framework/types.ts
export interface PanelSettingsContext<TConfig> {
  panelId: string;
  panelTitle: string;
  config: TConfig;
  setConfig: (next: TConfig | ((prev: TConfig) => TConfig)) => void;
  player: Player;
  topics: ReadonlyArray<TopicInfo>;
  resetPanel: () => void;
  duplicatePanel: () => void;
  closePanel: () => void;
  copyPanelId: () => void;
  extras?: Readonly<Record<string, unknown>>;
}